rday PR comment

+++ tests/unit/s2n_tls13_handshake_state_machine_test.c
@@ -326,13 +326,13 @@ int main(int argc, char **argv)
 
         EXPECT_SUCCESS(s2n_setup_handler_to_expect(SERVER_CHANGE_CIPHER_SPEC, S2N_CLIENT));
 
-        for (int i = 0; i < valid_tls13_handshakes_size; i++) {
+        for (size_t i = 0; i < valid_tls13_handshakes_size; i++) {
             int handshake = valid_tls13_handshakes[i];
 
             conn->handshake.handshake_type = handshake;
             conn->in_status = ENCRYPTED;
 
-            for (int j = 1; j < S2N_MAX_HANDSHAKE_LENGTH; j++) {
+            for (size_t j = 1; j < S2N_MAX_HANDSHAKE_LENGTH; j++) {
                 conn->handshake.message_number = j;
 
                 EXPECT_SUCCESS(s2n_test_write_header(&input, TLS_CHANGE_CIPHER_SPEC, 0));
@@ -403,13 +403,13 @@ int main(int argc, char **argv)
 
         EXPECT_SUCCESS(s2n_setup_handler_to_expect(CLIENT_CHANGE_CIPHER_SPEC, S2N_SERVER));
 
-        for (int i = 0; i < valid_tls13_handshakes_size; i++) {
+        for (size_t i = 0; i < valid_tls13_handshakes_size; i++) {
             int handshake = valid_tls13_handshakes[i];
 
             conn->handshake.handshake_type = handshake;
             conn->in_status = ENCRYPTED;
 
-            for (int j = 1; j < S2N_MAX_HANDSHAKE_LENGTH; j++) {
+            for (size_t j = 1; j < S2N_MAX_HANDSHAKE_LENGTH; j++) {
                 conn->handshake.message_number = j;
 
                 EXPECT_SUCCESS(s2n_test_write_header(&input, TLS_CHANGE_CIPHER_SPEC, 0));
